InternVQA_KVQ/requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
anykeystore==0.2
apex==0.9.10.dev0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
bitsandbytes==0.41.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
cmake==3.31.6
colorama==0.4.6
contourpy==1.3.1
cryptacular==1.6.2
cycler==0.12.1
decord==0.6.0
deepspeed==0.16.1
defusedxml==0.7.1
# dropout_layer_norm @ file:///mnt/data_nas/guanfb/VQA/single_modality_gfb_stage1/flash-attention-main/flash-attention-main/csrc/layer_norm
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.9
ffmpy==0.5.0
filelock==3.16.1
flash-attn==2.5.8
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
# fused_dense_lib @ file:///mnt/data_nas/guanfb/VQA/single_modality_gfb_stage1/flash-attention-main/flash-attention-main/csrc/fused_dense_lib
fvcore==0.1.5.post20221221
gradio==4.1.1
gradio_client==0.7.0
greenlet==3.1.1
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.29.1
hupper==1.12.1
icecream==2.1.4
idna==3.10
imageio==2.36.1
importlib_resources==6.5.2
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
latex2mathml==3.77.0
lazy_loader==0.4
linkify-it-py==2.0.3
lit==18.1.8
markdown-it-py==2.2.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.10.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
# Editable install with no version control (mplug_owl2==1.0.0)
# -e /mnt/data_nas/guanfb/VQA/Alibaba/Q_instruct/mPLUG-Owl-main/mPLUG-Owl2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.28.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opencv-python==4.8.0.76
orjson==3.10.15
packaging==24.2
pandas==2.0.3
PasteDeploy==3.1.0
pbkdf2==1.3
peft==0.4.0
Pillow==10.0.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
portalocker==3.0.0
propcache==0.3.0
protobuf==5.29.1
psutil==6.1.0
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.0
pyramid==2.0.2
pyramid-mailer==0.15.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
repoze.sendmail==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.2
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.45.3
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
timm==0.5.4
tokenizers==0.13.3
tomlkit==0.12.0
torch==2.0.1
torchaudio==2.5.1+cu118
torchvision==0.15.2
tqdm==4.67.1
transaction==5.0
transformers==4.31.0
translationstring==1.4
triton==2.0.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.34.0
velruse==1.1.1
venusian==3.1.1
wavedrom==2.0.3.post3
WebOb==1.8.9
websockets==11.0.3
WTForms==3.2.1
wtforms-recaptcha==0.3.2
yacs==0.1.8
yarl==1.18.3
zope.deprecation==5.0
zope.interface==7.2
zope.sqlalchemy==3.1
